+++ src-docs/src/views/form_layouts/inline_popover.js
@@ -112,6 +112,7 @@ export default () => {
     <div>
       <EuiPopover
         id="inlineFormPopover"
+        ownFocus
         button={button}
         isOpen={isPopoverOpen}
         closePopover={closePopover}>
@@ -120,6 +121,7 @@ export default () => {
       &emsp;
       <EuiPopover
         id="formPopover"
+        ownFocus
         button={button2}
         isOpen={isPopover2Open}
         closePopover={closePopover2}>