+++ src-docs/src/views/popover/popover_block.js
@@ -17,6 +17,7 @@ export default () => {
 
   return (
     <EuiPopover
+      ownFocus
       button={button}
       isOpen={isPopoverOpen}
       closePopover={closePopover}